Shuffling things around, taking code from existing rviz property
Looks like properties are the best way to look at things from rviz. Kind of a
rudimentary approximation of messages. Not sure of the best way to transfer data
back to the map once we want to change it, but that will come once we can view
the map data in the panel.