Now correctly refreshes on change, constructors filled in
Very inefficient - repopulates all properties when anything in the map changes.
Constructors all use delete. This should be changed later to not use pointers.

Currently has a segfault when a pose property is modified, think it's because
the current node is deleted and there is still something else referencing it.